post

+++ app/Http/Controllers/TasksController.php
@@ -18,6 +18,22 @@ class TasksController extends Controller
         return view('tasks.index', [
             'tasks' => $tasks,
         ]);
+
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザーを取得
+            $user = \Auth::user();
+            // ユーザーの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザーの投稿も取得するように変更しますが、現時点ではこのユーザーの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
     }
 
     /**
@@ -38,19 +54,19 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-         // バリデーション
+        // バリデーション
         $request->validate([
-            'status' => 'required|max:10',
-            'content' => 'required',
+            'content' => 'required|max:255',
         ]);
-         // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
 
-    // トップページ（タスク一覧）へリダイレクト
-        return redirect('/');
+        // 認証済みユーザー（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+        ]);
+
+        // 前のURLへリダイレクトさせる
+        return back();
+
     }
 
     /**
@@ -61,9 +77,17 @@ class TasksController extends Controller
           // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
+        // 関係するモデルの件数をロード
+        $user->loadRelationshipCounts();
+
+        // ユーザーの投稿一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+
     // タスク詳細ビューでそれを表示
         return view('tasks.show', [
-        'task' => $task,
+        'user' => $user,
+        'tasks' => $tasks
         ]);
     }
 
@@ -107,12 +131,20 @@ class TasksController extends Controller
      */
     public function destroy(string $id)
     {
-           // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // idの値で投稿を検索して取得
+        $task = Task::findOrFail($id);
+
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return back()
+                ->with('success','Delete Successful');
+        }
+
+        // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
+
     }
 }
